requirements.txt
Flask==3.0.0
flask-swagger-ui==4.11.1
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.2
gunicorn==21.2.0
python-dotenv==1.0.0